Zusatz gemacht Multiplikation
Hier sind:
- Zwei positive Zahlen addieren.
- Eine positive und eine negative Zahl addieren.
- 0 mit einer positiven Zahl addieren.
- Eine Zahl z. B. 10 und eine Zahl an der Grenze des Zahlenbereiches
(Integer.MAX_VALUE oder Integer.MIN_VALUE)
- Tests mit Exceptions…
mit der Multiplikation

+++ src/main/java/ch/bbw/pr/savecalculator/SaveCalculator.java
@@ -38,7 +38,7 @@ public class SaveCalculator {
   }
   
 }
-  public double multiplikation(int value1, int value2) {
+  public int multiplikation(int value1, int value2) {
     long value = (long) value1 * (long) value2;
     if ((value > Integer.MAX_VALUE) || (value < Integer.MIN_VALUE)){
        throw new ArithmeticException();
